backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
PyJWT
python-multipart
pydantic
pydantic-settings
celery
redis
httpx
requests
python-dotenv
pytest
pytest-cov
pytest-asyncio
beautifulsoup4
numpy
psycopg2-binary
aiohttp
Pillow
urllib3
stripe
bcrypt
